+++ asdf/_core/_converters/complex.py
@@ -13,7 +13,6 @@ _REPLACEMENTS = {
 
 class ComplexConverter(Converter):
     tags = ["tag:stsci.edu:asdf/core/complex-1.0.0"]
-
     types = [*list(util._iter_subclasses(np.complexfloating)), complex]
 
     def to_yaml_tree(self, obj, tag, ctx):