+++ asdf/_tests/test_extension.py
@@ -5,6 +5,7 @@ import pytest
 from packaging.specifiers import SpecifierSet
 from yaml.representer import RepresenterError
 
+import asdf
 from asdf import AsdfFile, config_context
 from asdf.exceptions import AsdfManifestURIMismatchWarning, AsdfWarning, ValidationError
 from asdf.extension import (
@@ -899,7 +900,7 @@ def test_warning_or_error_for_default_select_tag(is_subclass, indirect):
             config.add_extension(extension)
 
 
-def test_reference_cycle():
+def test_reference_cycle(tmp_path, with_lazy_tree):
     class FractionWithInverse(fractions.Fraction):
         def __init__(self, *args, **kwargs):
             self._inverse = None
@@ -940,9 +941,11 @@ def test_reference_cycle():
         f2 = FractionWithInverse(5, 3)
         f1.inverse = f2
         f2.inverse = f1
-
-        read_f1 = roundtrip_object(f1)
-        assert read_f1.inverse.inverse is read_f1
+        fn = tmp_path / "test.asdf"
+        asdf.AsdfFile({"obj": f1}).write_to(fn)
+        with asdf.open(fn) as af:
+            read_f1 = af["obj"]
+            assert read_f1.inverse.inverse is read_f1
 
 
 def test_manifest_uri_id_mismatch_warning(tmp_path):